refactor(App) : Editor 컴포넌트 파일 분리
- Editor를 별도의 컴포넌트 파일로 분리하였습니다.
